requirement.txt
Django==4.2.4
Pillow==10.0.0
djangorestframework==3.14.0